move glm test in test_model
